Make Speedrun create a phantom
This makes logical sense for the joker. Based on a suggestion from here: https://discord.com/channels/1226193436521267223/1396398881054527498

The idea is that it *should* alert you that someone has picked up a speedrun, so you can know if you need to play fast or not, which helps balance it. As an example, if someone is playing carefully in a small blind while someone else is playing quick, they could see the speedrun first, and then the careful player gets punished through no fault of their own.