Copy to Clipboard Button on Lobby Screen
This adds a copy to clipboard button to the lobby screen, to prevent having to click 3 times to copy to clipboard, changing it to just 1. I don't really know how to utilize the localize function as any time I tried to the game would crash, so no use of that in here, but otherwise works.